# Small perf optimizations to statement preparation